extract nav careers
